style: move mixins to new fishing mixin package
