--- src/client/resources/mixins/skyocean.bobberfix.mixins.json
@@ -1,8 +1,0 @@
-{
-  "required": false,
-  "package": "me.owdding.skyocean.mixins.features.bobberfix",
-  "parent": "skyocean.client.mixins.json",
-  "client": [
-    "FishingHookMixin"
-  ]
-}
