+++ src/main/resources/fabric.mod.json
@@ -23,8 +23,7 @@
     "skyocean.client.mixins.json",
     "mixins/skyocean.hidearmour.mixins.json",
     "mixins/skyocean.hidelightning.mixins.json",
-    "mixins/skyocean.lavareplacement.mixins.json",
-    "mixins/skyocean.bobberfix.mixins.json"
+    "mixins/skyocean.fishing.mixins.json"
   ],
   "accessWidener": "skyocean.accesswidener",
   "depends": {
